Completed the prototype for DFParse which ends up pulling all values for an associated ID in page chronological order, which means that the entire log is reversed. Need to rewrite logic such that each <p> is stuffed in an array and then reverse written into the file. Will probably make a learner of some kind to extract statistics with data science in conjunction with this script later on.
Duel is just about done! Finally added in promotion battle logic to strip leaders as well as the team. It could be more efficient, but since it only adds ~5-10 minutes onto the runtime I won't be improving this aspect any further. killYourself() has been completed and now successfully allows for multiple streaks to be executed per script. The only drawback is the 'smurf' logic for the first 15-20 matches that ignores the maximum level limit value. This has killed me multiple times immediately on execution and is designed for use with extended streaking. To counter this, another user input flag has been included so that this can be turned off if the user runs the script more than once a day.